Fix race-condition in Selection.on_change.
An owner-change triggered by an external application could interfere with
the locking in Selection by triggering the code path meant to clear the
recursion inhibiting lock after using Selection.set. If another thread
acquired the lock in the meantime, the external event would break the lock;
we would subsequently call release() on an already unlocked lock.

The change tries to avoid further race conditions by maintaining a deque,
which gives the information required to avoid infinite recursion, and
abstracts away from the sequence in which owner-change events happen.